requirements.txt
# HealthBot - Multilingual Public Health Chatbot
# Python 3.10+

# API & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
requests>=2.31.0

# NLP & Embeddings (optional - fallback to random if not installed)
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
langdetect>=1.0.9
indic-transliteration>=2.2.0

# ML
scikit-learn>=1.3.2
numpy>=1.26.0
pandas>=2.1.0

# Knowledge Base (in-memory / optional Neo4j)
py2neo>=2021.2.3

# Auth
sqlalchemy>=2.0.0
passlib[argon2]>=1.7.4
argon2-cffi>=21.3.0
python-jose[cryptography]>=3.3.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
